organize into components
