+++ web/src/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import WikipediaBrowser from "../components/WikipediaBrowser/WikipediaBrowser";
+import WikipediaBrowser from "@/components/WikipediaBrowser/WikipediaBrowser";
 
 
 const Home: React.FC = () => {
