+++ web/src/components/Pane/types.ts
@@ -1,0 +1,6 @@
+export default interface Pane {
+  title: string;
+  content: string;
+  isSearchResult: boolean;
+  width: number;
+}
